Testing preconditions to improve performance.
Test the idea that by prepending the set of syscalls/events to the
beginning of the big or'd expression as an initial "evt/syscall
in (...) and...", we can improve performance.

The idea is that the big expression will only be evaluated for those
events/syscalls actually used in the rules of the expression.

This commit entirely hardcodes the set of syscalls/events by examining
the current falco_rules.yaml file and seeing which syscalls/events are
used. If this pans out, we'll replace this with an automatic step.

This commit also adds syscalls/event conditions as much as possible to
rules that do not have them.